[UDT-92] refactor: profile 타입 폴더 이름 변경 및 타입 수정
