+++ src/hooks/profile/useInfiniteFeedbacks.ts
@@ -3,8 +3,8 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 import type {
   FeedbackQueryParams,
   GetFeedbackContentsResponse,
-} from '@type/mypage/RecommendedContent';
-import { getFeedbackContents } from '@/lib/apis/profile/feedbackService';
+} from '@type/profile/RecommendedContent';
+import { getFeedbackContents } from '@lib/apis/profile/feedbackService';
 
 export const useInfiniteFeedbacks = (
   baseParams: Omit<FeedbackQueryParams, 'cursor'>,
