+++ src/lib/apis/profile/feedbackService.ts
@@ -1,7 +1,7 @@
 import {
   FeedbackQueryParams,
   GetFeedbackContentsResponse,
-} from '@type/mypage/RecommendedContent';
+} from '@type/profile/RecommendedContent';
 import axiosInstance from '../axiosInstance';
 
 // [GET] /api/users/me/feedbacks → 좋아요/싫어요 피드백 콘텐츠 조회
