+++ src/lib/apis/profile/recommendService.ts
@@ -1,4 +1,4 @@
-import { RecommendedContent } from '@type/mypage/RecommendedContent';
+import { RecommendedContent } from '@type/profile/RecommendedContent';
 import axiosInstance from '../axiosInstance';
 
 // [GET] /api/users/me/curated/contents 사용자의 큐레이션 콘텐츠 목록 조회
