--- src/types/UserInfo.ts
@@ -1,7 +1,0 @@
-export interface UserInfo {
-  name: string;
-  email: string;
-  platforms: string[];
-  genres: string[];
-  profileImageUrl?: string;
-}
